+++ packages/applications/document-builder/.storybook/preview.tsx
@@ -3,21 +3,18 @@ import React from 'react';
 import { Font, PDFViewer } from '@react-pdf/renderer';
 import type { Preview } from '@storybook/react';
 
-const assetsFolder = path.resolve('../src/assets');
-const fontsFolderPath = path.resolve(assetsFolder, 'fonts');
-
 Font.register({
   family: 'Arimo',
   fonts: [
     {
-      src: path.join(fontsFolderPath, '/arimo/Arimo-Regular.ttf'),
+      src: '/fonts/arimo/Arimo-Regular.ttf',
     },
     {
-      src: path.join(fontsFolderPath, '/arimo/Arimo-Bold.ttf'),
+      src: '/fonts/arimo/Arimo-Bold.ttf',
       fontWeight: 'bold',
     },
     {
-      src: path.join(fontsFolderPath, '/arimo/Arimo-Italic.ttf'),
+      src: '/fonts/arimo/Arimo-Italic.ttf',
       fontStyle: 'italic',
     },
   ],
@@ -27,7 +24,7 @@ const preview: Preview = {
   decorators: [
     (Story) => {
       return (
-        <PDFViewer width="100%" height="900px">
+        <PDFViewer width="100%" height="1200px">
           <Story />
         </PDFViewer>
       );
