+++ packages/applications/document-builder/src/abandon/accordAbandonAvecRecandidature/buildDocument.ts
@@ -1,11 +1,31 @@
-import ReactPDF from '@react-pdf/renderer';
+import path from 'node:path';
+
+import ReactPDF, { Font } from '@react-pdf/renderer';
 
 import { mapToReadableStream } from '../../mapToReadableStream';
+import { fontsFolderPath } from '../../assets';
 
 import {
   RéponseAbandonAvecRecandidature,
   RéponseAbandonAvecRecandidatureProps,
 } from './RéponseAbandonAvecRecandidature';
+
+Font.register({
+  family: 'Arimo',
+  fonts: [
+    {
+      src: path.join(fontsFolderPath, '/arimo/Arimo-Regular.ttf'),
+    },
+    {
+      src: path.join(fontsFolderPath, '/arimo/Arimo-Bold.ttf'),
+      fontWeight: 'bold',
+    },
+    {
+      src: path.join(fontsFolderPath, '/arimo/Arimo-Italic.ttf'),
+      fontStyle: 'italic',
+    },
+  ],
+});
 
 export type GénérerRéponseAccordAbandonAvecRecandidaturePort = (
   données: DonnéesDocument,
