+++ packages/applications/document-builder/src/assets.ts
@@ -1,5 +1,5 @@
 import path from 'node:path';
 
-const assetsFolderPath = path.resolve(__dirname, 'assets');
+export const assetsFolderPath = path.resolve(__dirname, 'assets');
 export const fontsFolderPath = path.resolve(assetsFolderPath, 'fonts');
 export const imagesFolderPath = path.resolve(assetsFolderPath, 'images');
