+++ packages/applications/document-builder/src/candidature/attestation/ppe2.v1/makeCertificate.tsx
@@ -15,6 +15,7 @@ const makeCertificate = (
   const { content, footnotes } = project.isClasse
     ? buildLauréat({ project })
     : buildElimine({ project });
+
   return (
     <Certificate project={project} validateur={validateur} footnotes={footnotes}>
       {content}
